test: consolidate test files, delete factories.py, rename test_container
- Move DbFactory from factories.py into helpers.py, migrate all imports
- Delete tests/factories.py (86 LOC)
- Rename test_container.py → test_services.py (tests services.py)
- Merge test_backend_core.py INTO test_storage.py (99 tests combined)
- Merge test_search_core.py INTO test_search_index.py (90 tests combined)
- Merge test_cli_integration.py INTO test_cli_commands.py (1236 LOC)
- Add .metadata() method to ConversationBuilder

All 2206 tests pass, 0 failures.
