+++ tests/test_core.py
@@ -782,7 +782,7 @@ def test_repository(test_db):
     backend = SQLiteBackend(db_path=test_db)
     repo = ConversationRepository(backend=backend)
 
-    from tests.factories import DbFactory
+    from tests.helpers import DbFactory
 
     factory = DbFactory(test_db)
     factory.create_conversation(
@@ -804,7 +804,7 @@ def test_repository(test_db):
 
 def test_repository_get_includes_attachment_conversation_id(test_db):
     """ConversationRepository.get_eager() returns attachments with conversation_id field."""
-    from tests.factories import DbFactory
+    from tests.helpers import DbFactory
 
     factory = DbFactory(test_db)
     factory.create_conversation(
@@ -842,7 +842,7 @@ def test_repository_get_includes_attachment_conversation_id(test_db):
 
 def test_repository_get_with_multiple_attachments(test_db):
     """get_eager() correctly groups multiple attachments per message."""
-    from tests.factories import DbFactory
+    from tests.helpers import DbFactory
 
     factory = DbFactory(test_db)
     factory.create_conversation(
@@ -888,7 +888,7 @@ def test_repository_get_with_multiple_attachments(test_db):
 
 def test_repository_get_attachment_metadata_decoded(test_db):
     """Attachment provider_meta JSON is properly decoded."""
-    from tests.factories import DbFactory
+    from tests.helpers import DbFactory
 
     factory = DbFactory(test_db)
     meta = {"original_name": "photo.png", "source": "upload"}
